Move Constants.kt -> spell/ElementType.kt
