+++ core/src/main/kotlin/no/ntnu/beardblaster/spell/ElementType.kt
@@ -1,4 +1,4 @@
-package no.ntnu.beardblaster
+package no.ntnu.beardblaster.spell
 
 enum class ElementType {
     Fire,
