+++ core/src/main/kotlin/no/ntnu/beardblaster/spell/SpellBar.kt
@@ -5,7 +5,6 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table
 import ktx.actors.onClick
 import ktx.log.logger
 import ktx.scene2d.*
-import no.ntnu.beardblaster.ElementType
 import no.ntnu.beardblaster.commons.spell.Element
 import no.ntnu.beardblaster.ui.Image
 import no.ntnu.beardblaster.ui.get
