+++ core/src/main/kotlin/no/ntnu/beardblaster/ui/Skin.kt
@@ -12,7 +12,7 @@ import com.badlogic.gdx.scenes.scene2d.utils.SpriteDrawable
 import ktx.assets.async.AssetStorage
 import ktx.scene2d.Scene2DSkin
 import ktx.style.*
-import no.ntnu.beardblaster.ElementType
+import no.ntnu.beardblaster.spell.ElementType
 import no.ntnu.beardblaster.assets.AtlasAsset
 import no.ntnu.beardblaster.assets.FontAsset
 import no.ntnu.beardblaster.assets.get
